requirements.txt
google-genai
python-dotenv
tqdm
langchain-experimental
langchain-huggingface
sentence-transformers
langchain
torch